feat: Added bioformats support (#16)
* chore: Added bioformats

* chore: Added JVM

* chore: Test compile

* chore: First running bioformats loader

* chore: Fixed image read

* chore: Fixed bioformats loader

* chore: Moved AI models

* chore: Some fixed in reporting and bioformats

* chore: Added some more file format support

* chore: Smaller fixes

* chore: Removed debugging from action

+++ src/backend/image_reader/jpg/image_loader_jpg.cpp
@@ -25,10 +25,10 @@ auto JpgLoader::readImageMeta(const std::string &filename) -> std::string
 }
 auto JpgLoader::getImageProperties(const std::string &filename) -> ImageProperties
 {
-  auto image = cv::imread(filename, cv::IMREAD_COLOR);
-  int width  = image.cols;
-  int height = image.rows;
-  int size   = width * height;
+  auto image   = cv::imread(filename, cv::IMREAD_COLOR);
+  int width    = image.cols;
+  int height   = image.rows;
+  int64_t size = width * height;
 
   return ImageProperties{.imageSize = size, .tileSize = size, .nrOfTiles = 1, .nrOfDocuments = 1};
 }
